requirements.txt
redis
flask
flask-socketio
python-socketio
requests
websocket-client
wxpython

